+++ src/yasoo/__init__.py
@@ -1,5 +1,3 @@
-from datetime import datetime
-
 from .deserialization import Deserializer
 from .serialization import Serializer
 
@@ -12,13 +10,3 @@ _default_deserializer = Deserializer()
 deserialize = _default_deserializer.deserialize
 deserializer = _default_deserializer.register()
 deserializer_of = _default_deserializer.register
-
-
-@serializer_of(datetime)
-def _serialize_datetime(d: datetime) -> dict:
-    return {"time": d.timestamp()}
-
-
-@deserializer_of(datetime)
-def _deserialize_datetime(d: dict) -> datetime:
-    return datetime.fromtimestamp(d["time"])